Cookie Store API: add branch for empty string path https://bugs.webkit.org/show_bug.cgi?id=297268
Reviewed by Rupin Mittal.

Implement https://github.com/whatwg/cookiestore/pull/283 with tests
upstreamed at https://github.com/web-platform-tests/wpt/pull/54264

Canonical link: https://commits.webkit.org/298681@main
