+++ LayoutTests/imported/w3c/web-platform-tests/cookiestore/cookieStore_set_path.https.window-expected.txt
@@ -1,0 +1,4 @@
+
+PASS CookieListItem - cookieStore.set with empty string path defaults to current URL
+PASS CookieListItem - cookieStore.set with empty string path defaults to current URL with __host- prefix
+
