+++ LayoutTests/imported/w3c/web-platform-tests/cookiestore/cookieStore_set_path.https.window.html
@@ -1,0 +1,1 @@
+<!-- This file is required for WebKit test infrastructure to run the templated test -->
