+++ Source/WebCore/platform/network/Cookie.cpp
@@ -42,6 +42,25 @@ unsigned Cookie::hash() const
     return StringHash::hash(name) + StringHash::hash(domain) + StringHash::hash(path) + secure;
 }
 #endif
-    
+
+namespace CookieUtil {
+
+String defaultPathForURL(const URL& url)
+{
+    // Algorithm to generate the default path is outlined in https://tools.ietf.org/html/rfc6265#section-5.1.4
+
+    String path = url.path().toString();
+    if (path.isEmpty() || !path.startsWith('/'))
+        return "/"_s;
+
+    auto lastSlashPosition = path.reverseFind('/');
+    if (!lastSlashPosition)
+        return "/"_s;
+
+    return path.left(lastSlashPosition);
+}
+
+} // namespace CookieUtil
+
 } // namespace WebCore
